toppartitions: Improve cardinality precision and pass list_size to gather method
In the next patches it'll be added support for returning toppartitions results per shard.
Currently, this is not possible because even though the toppartitions are already calculated
per shard, we're trimming the results in the gather method by passing the capacity as
an argument instead of the actual list size.
This patch fixes this by passing the list_size to the gather method while making sure that
the topk object which is receiving the entries has a capacity of `list_size * smp::count`.
It also improves the precision of the final calculated cardinality by summing the
cardinalities calculated in each shard instead of relying on the size of the results vector.
The changes made here don't affect the size of the results returned by the toppartitions
API because the results are trimmed to the list_size in the api method.