tests: Mitigate flaky tests annoyance
The test.py has a facility to soften the impact from flaky tests -- if
a test is marked as flaky in the suite.yaml file, it can be restarted
several times if it fails. This functionalty has several drawbacks.

 First: flaky test can fail more than "several" times in a row
 Second: no tests are marked as flaky in current master
 Third: there still can be flaky tests we're unaware of

Having said that, here's an improvement -- the --run-until-it-pass
option that keeps restarting any test (not only previously identified
and marked as flaky) until it passes. This makes sure no flaky (both
known and unknown) tests can break through the existing protection.

While at it, to make sure this new option works as long as the existing
flaky test restart loop -- add a simple flaky test and mark it as such.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>