sstables_loader: Make non-concurrent loading more verbose
Current load-new-sstables API allows limited parallelizm. If hashes
table name and by that hash value identifies the shard that will govern
the loading. One shard is allowed to serve at most one request.

To enforce that limitation, sstables_loader instance have bool flag on
board that's set before loading and unset after it passes or fails, and
if a new attempt is made on this shard an exception is thrown.

The patch extends the boolean to contain ks.cf pair from which sstables
are being loaded on current shard, so that the API caller could get more
info upon "try again later" exception.

Also, while at it, make the boolean reset more RAIIsh.

Signed-off-by: Pavel Emelyanov <xemul@scylladb.com>