test: network_topology_strategy_test: use std::lerp
std::lerp() interpolate a value in the given range into another,
and we don't need to worry about the overflow, so it is a perfect
fit to our needs. as `d2t()` scales a float point number in the
range of [0, 1] to the a token represented with an unsigned long.
so, let's use `std::lerp()` to replace the shift dance.

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>