test/cql-pytest: test out-of-range bigint from fromJson()
user could insert out-of-range numbers using a float point
number via "fromJson()" as bigint. and we should reject them.
please note this is not exactly same issue as #10137, where we
can barely do nothing other than improving RapidJson or use
another library for parsing JSON so that it does not lose the
accuracy when interpreting the float point numbers. while
in this case, the accuracy is not lost because of RapidJson.
instead, we're obliged to prevent user from inserting out-of-range
numbers using fromJson() even if these numbers are very close
to the boundaries of the range in question and these numbers
are not rounded by RapidJson.

Refs #13077

Signed-off-by: Kefu Chai <kefu.chai@scylladb.com>