sstables: make_entry_descriptor(): use s/std::regex/boost::regex/
As demonstrated countless times, std::regex's practice of using
recursion for pattern matching can result in stack overflow, especially
on AARCH64. The most recent incident happened after merging #13075,
which (indirectly) uses `make_entry_descriptor()` to test whether a
certain path is a valid scylla table path in a trial-and-error manner.
This resulted in stacks blowing up in AARCH64.
To prevent this, use the already tried and tested method of switching
from `std::regex` to `boost::regex`.

Refs: #13404