sstables: Don't store token data into summary's memory pool
summary has a memory pool, which is implemented as a set of contiguous
buffer of exponentially increasing size, with the max size of 128k.

This pool served for both storing keys of summary entries and their
respective tokens. The summary entry itself just stores a string_view,
which points to the actual data in the memory pool.

Since this series 31593e1451f4, which removed token_view, summary_entry
stores the actual token, not just the view.

Therefore, memory is being wasted, as SSTable loader / writer is
unnecessarily storing the token data into the pool.

With 11k summary entries, the footprint drops from 756004 to 624932.
A 18% reduction. Of course, the reduction depends on factors like key
size, where the key size can outweigh significantly this waste.

Signed-off-by: Raphael S. Carvalho <raphaelsc@scylladb.com>